third_party/WebKit/LayoutTests/svg/custom/acid3-test-77-expected.txt

Acid3 test 77, checking text metric functionality

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.getNumberOfChars() is 3
PASS text.getComputedTextLength() is 4711+42+23
PASS text.getSubStringLength(0,1) is 42
PASS text.getSubStringLength(0,2) is 42+23
PASS text.getSubStringLength(1,1) is 23
PASS text.getSubStringLength(1,0) is 0
PASS text.getSubStringLength(1, 3) is 4734
PASS text.getSubStringLength(0, 4) is 4776
PASS text.getSubStringLength(3, 0) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (3) is greater than or equal to the maximum bound (3)..
PASS text.getSubStringLength(-17, 20) threw exception IndexSizeError: Failed to execute 'getSubStringLength' on 'SVGTextContentElement': The charnum provided (4294967279) is greater than the maximum bound (3)..
PASS text.getStartPositionOfChar(0).x is 0
PASS text.getStartPositionOfChar(1).x is 42
PASS text.getStartPositionOfChar(2).x is 42+23
PASS text.getStartPositionOfChar(0).y is 4000
PASS text.getEndPositionOfChar(-1) threw exception IndexSizeError: Failed to execute 'getEndPositionOfChar' on 'SVGTextContentElement': The charnum provided (4294967295) is greater than the maximum bound (3)..
PASS text.getEndPositionOfChar(4) threw exception IndexSizeError: Failed to execute 'getEndPositionOfChar' on 'SVGTextContentElement': The charnum provided (4) is greater than the maximum bound (3)..
PASS text.getEndPositionOfChar(0).x is 42
PASS text.getEndPositionOfChar(1).x is 42+23
PASS text.getEndPositionOfChar(2).x is 42+23+4711
PASS text.getEndPositionOfChar(-17) threw exception IndexSizeError: Failed to execute 'getEndPositionOfChar' on 'SVGTextContentElement': The charnum provided (4294967279) is greater than the maximum bound (3)..
PASS text.getEndPositionOfChar(4) threw exception IndexSizeError: Failed to execute 'getEndPositionOfChar' on 'SVGTextContentElement': The charnum provided (4) is greater than the maximum bound (3)..
PASS successfullyParsed is true

TEST COMPLETE
